backend/JobSearchAgent/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
bedrock-agentcore
boto3>=1.40.0
botocore>=1.40.0
requests
